WIP upgrading to 1.19.4

+++ imm_ptl_core/src/main/java/qouteall/imm_ptl/core/portal/global_portals/WorldWrappingPortal.java
@@ -140,10 +140,10 @@ public class WorldWrappingPortal extends GlobalTrackedPortal {
             
             return new IntBox(
                 new BlockPos(
-                    Math.round(floatBox.minX), McHelper.getMinY(world), Math.round(floatBox.minZ)
+                    (int) Math.round(floatBox.minX), McHelper.getMinY(world), (int) Math.round(floatBox.minZ)
                 ),
                 new BlockPos(
-                    Math.round(floatBox.maxX) - 1, McHelper.getMaxYExclusive(world), Math.round(floatBox.maxZ) - 1
+                    (int) Math.round(floatBox.maxX) - 1, McHelper.getMaxYExclusive(world), (int) Math.round(floatBox.maxZ) - 1
                 )
             );
         }
@@ -158,10 +158,10 @@ public class WorldWrappingPortal extends GlobalTrackedPortal {
             
             return new IntBox(
                 new BlockPos(
-                    Math.round(floatBox.minX) - 1, McHelper.getMinY(world), Math.round(floatBox.minZ) - 1
+                    (int) Math.round(floatBox.minX) - 1, McHelper.getMinY(world), (int) Math.round(floatBox.minZ) - 1
                 ),
                 new BlockPos(
-                    Math.round(floatBox.maxX), McHelper.getMaxYExclusive(world), Math.round(floatBox.maxZ)
+                    (int) Math.round(floatBox.maxX), McHelper.getMaxYExclusive(world), (int) Math.round(floatBox.maxZ)
                 )
             );
         }
